Final Development Commit
This commit marks the end of development of Fortnite Gun Statistics. All further commits will only be to add new weapons and changes that come to Fortnite. Other than that, however, the project is finished.

I want to look back at a few important dates that highlighted the development stages of the project.

October 4th, 2018 - Development of the website officially begins.

October 5th, 2018 to December 3rd, 2018 - Initial development of the website continues.

December 4th, 2018 - The website officially launches to the public internet.

December 5th, 2018 to May 29th, 2019 - Post-launch development of the website continues.

May 30th, 2019 - The project is officially finished with development and any new changes will simply be new weapons and changes added to Fortnite.

It has been a fun project to work on, but it is time to move on to bigger and better thing.